requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
crispy-tailwind==1.0.3
cryptography==42.0.7
deprecation==2.1.0
diff-match-patch==20230430
Django==4.2.13
django-allauth==0.62.1
django-appconf==1.0.6
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-countries==7.6.1
django-countries-with-calling-codes==0.1.1
django-crispy-forms==2.1
django-easy-audit==1.3.6
django-email-verification==0.3.3
django-environ==0.11.2
django-formtools==2.3
django-htmx==1.17.3
django-import-export==4.0.1
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-resized==1.0.2
django-select2==8.1.2
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
maxminddb==2.6.1
multidict==6.0.5
mysqlclient==2.2.4
numpy==1.24.4
oauthlib==3.2.2
packaging==24.0
pdfkit==1.0.0
phonenumbers==8.13.36
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python-magic-bin==0.4.14
pytz==2023.3
PyYAML==6.0.1
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.1
validators==0.28.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
